refactor: Move llama.js to oracle folder
